Renamed logout.js to logged-in-menu.js to reflect component name.
